+++ src/main/java/com/acd/researchrepo/repository/RefreshTokenRepository.java
@@ -4,6 +4,6 @@ import com.acd.researchrepo.model.RefreshToken;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Integer> {
 
 }
